[iOS][updates] Migrate tests to Swift Testing (3/3)
